+++ modules/kubernetes-cluster/src/variables.tf
@@ -21,6 +21,17 @@ variable "identifier" {
   type = string
 }
 
+variable "identity" {
+  type = object({
+    type         = string
+    identity_ids = optional(list(string))
+  })
+  default = {
+    type         = "SystemAssigned"
+    identity_ids = []
+  }
+}
+
 variable "kubernetes_version" {
   type = string
 }