+++ modules/kubernetes-cluster/CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 ## 4.0
-* Updated the AzureRM Terraform Provider from 3.x to 4.x.
+* Updated the required Terraform version from `1.5` to `1.11`.
+* Updated the AzureRM Terraform provider version from `3.x` to `4.x`.
 
 ## 3.0
 * Added `os_disk_size_gb`, `os_disk_type` and `os_sku` variables to enable greater default node pool customisation.